% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NNETARCH-package.R
\docType{package}
\name{NNETARCH-package}
\alias{NNETARCH}
\alias{NNETARCH-package}
\title{NNETARCH: Neural Network Autoregressive Conditional Heteroskedasticity for Time Series Forecasting}
\description{
Implements a hybrid forecasting model combining nonlinear autoregression via neural networks (NNETAR) with volatility modeling using either GARCH or neural network-based residual modeling. This package extends the 'forecast' package functionality to provide variance-adaptive forecasting models for financial and economic time series.
}
\author{
\strong{Maintainer}: Rafael Montano \email{rafael.montano@sydney.edu.au}

}
