% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnetarch.R
\name{forecast.nnetarch}
\alias{forecast.nnetarch}
\title{Forecast Method for NNETARCH Objects}
\usage{
\method{forecast}{nnetarch}(object, h = 10, level = c(80, 95), ...)
}
\arguments{
\item{object}{An object of class \code{nnetarch}.}

\item{h}{Number of steps ahead for forecasting.}

\item{level}{Confidence levels for prediction intervals (numeric vector). Default is \code{c(80, 95)}.}

\item{...}{Additional arguments passed to the underlying \code{forecast()} method.}
}
\value{
An object of class \code{forecast}.
}
\description{
Generates forecasts from an object of class \code{nnetarch}.
}
